+++ include/heyoka/callable.hpp
@@ -9,107 +9,170 @@
 #ifndef HEYOKA_CALLABLE_HPP
 #define HEYOKA_CALLABLE_HPP
 
+#include <concepts>
 #include <functional>
-#include <memory>
 #include <type_traits>
-#include <typeindex>
-#include <typeinfo>
 #include <utility>
 
 #include <heyoka/config.hpp>
+#include <heyoka/detail/tanuki.hpp>
 #include <heyoka/detail/type_traits.hpp>
 #include <heyoka/detail/visibility.hpp>
 #include <heyoka/s11n.hpp>
 
-// NOTE: here we implement a stripped-down version of std::function
-// on top of classic OOP polymorphism.
-//
-// The reason for this (as opposed to using std::function directly)
-// is principally because we can make our implementation serialisable
-// via Boost.Serialization, whereas std::function cannot be supported
-// by Boost. We need a serialisable function wrapper in order to be
-// able to serialise the callbacks of the events, whose serialisation,
-// in turn, is needed in the serialisation of the integrator objects.
+#if defined(__GNUC__)
+
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wmissing-field-initializers"
+
+#endif
 
 HEYOKA_BEGIN_NAMESPACE
+
+// Fwd declaration of the detector
+// for any callable object.
+template <typename>
+struct is_any_callable;
 
 namespace detail
 {
 
+// Declaration of the callable interface template.
+template <typename, typename, typename, typename...>
+struct HEYOKA_DLL_PUBLIC_INLINE_CLASS callable_iface {
+};
+
+// Declaration of the callable interface.
 template <typename R, typename... Args>
-struct HEYOKA_DLL_PUBLIC_INLINE_CLASS callable_inner_base {
-    callable_inner_base() = default;
-    callable_inner_base(const callable_inner_base &) = delete;
-    callable_inner_base(callable_inner_base &&) = delete;
-    callable_inner_base &operator=(const callable_inner_base &) = delete;
-    callable_inner_base &operator=(callable_inner_base &&) = delete;
-    virtual ~callable_inner_base() = default;
+// NOLINTNEXTLINE(cppcoreguidelines-special-member-functions,hicpp-special-member-functions)
+struct HEYOKA_DLL_PUBLIC_INLINE_CLASS callable_iface<void, void, R, Args...> {
+    virtual ~callable_iface() = default;
+    virtual R operator()(Args... args) = 0;
+    virtual explicit operator bool() const noexcept = 0;
+};
 
-    virtual std::unique_ptr<callable_inner_base> clone() const = 0;
-
-    virtual R operator()(Args...) const = 0;
-
-    [[nodiscard]] virtual std::type_index get_type_index() const = 0;
-
-private:
-    // Serialization (empty, no data members).
-    friend class boost::serialization::access;
-    template <typename Archive>
-    void serialize(Archive &, unsigned)
+// Implementation of the callable interface for
+// invocable objects.
+template <typename Holder, typename T, typename R, typename... Args>
+    requires std::is_invocable_r_v<R, std::remove_reference_t<std::unwrap_reference_t<T>> &, Args...>
+                 // NOTE: also require copy constructability like
+                 // std::function does.
+                 && std::copy_constructible<T>
+struct HEYOKA_DLL_PUBLIC_INLINE_CLASS callable_iface<Holder, T, R, Args...>
+    : virtual callable_iface<void, void, R, Args...>, tanuki::iface_impl_helper<Holder, T, callable_iface, R, Args...> {
+    R operator()(Args... args) final
     {
+        using unrefT = std::remove_reference_t<std::unwrap_reference_t<T>>;
+
+        if constexpr (std::is_pointer_v<unrefT> || std::is_member_pointer_v<unrefT>) {
+            if (this->value() == nullptr) {
+                throw std::bad_function_call{};
+            }
+        }
+
+        // NOTE: if this->value() is an empty std::function or callable,
+        // the std::bad_function_call exception will be raised
+        // by the invocation.
+
+        if constexpr (std::is_same_v<R, void>) {
+            static_cast<void>(std::invoke(this->value(), std::forward<Args>(args)...));
+        } else {
+            return std::invoke(this->value(), std::forward<Args>(args)...);
+        }
+    }
+    explicit operator bool() const noexcept final
+    {
+        using unrefT = std::remove_reference_t<std::unwrap_reference_t<T>>;
+
+        if constexpr (std::is_pointer_v<unrefT> || std::is_member_pointer_v<unrefT>) {
+            return this->value() != nullptr;
+        } else if constexpr (is_any_callable<unrefT>::value || is_any_std_func_v<unrefT>) {
+            return static_cast<bool>(this->value());
+        } else {
+            return true;
+        }
     }
 };
 
-template <typename T, typename R, typename... Args>
-struct HEYOKA_DLL_PUBLIC_INLINE_CLASS callable_inner final : callable_inner_base<R, Args...> {
-    T m_value;
+// Implementation of the reference interface.
+template <typename Wrap, typename R, typename... Args>
+struct HEYOKA_DLL_PUBLIC_INLINE_CLASS callable_ref_iface_impl {
+    using result_type = R;
 
-    // We just need the def ctor, delete everything else.
-    callable_inner() = default;
-    callable_inner(const callable_inner &) = delete;
-    callable_inner(callable_inner &&) = delete;
-    callable_inner &operator=(const callable_inner &) = delete;
-    callable_inner &operator=(callable_inner &&) = delete;
-    ~callable_inner() final = default;
-
-    // Constructors from T (copy and move variants).
-    explicit callable_inner(const T &x) : m_value(x) {}
-    explicit callable_inner(T &&x) : m_value(std::move(x)) {}
-
-    // The clone method, used in the copy constructor.
-    std::unique_ptr<callable_inner_base<R, Args...>> clone() const final
+    template <typename JustWrap = Wrap, typename... FArgs>
+    auto operator()(FArgs &&...fargs)
+        -> decltype(iface_ptr(*static_cast<JustWrap *>(this))->operator()(std::forward<FArgs>(fargs)...))
     {
-        return std::make_unique<callable_inner>(m_value);
+        if (is_invalid(*static_cast<Wrap *>(this))) {
+            throw std::bad_function_call{};
+        }
+
+        return iface_ptr(*static_cast<Wrap *>(this))->operator()(std::forward<FArgs>(fargs)...);
     }
 
-    R operator()(Args... args) const final
+    explicit operator bool() const noexcept
     {
-        return m_value(std::forward<Args>(args)...);
+        if (is_invalid(*static_cast<const Wrap *>(this))) {
+            return false;
+        } else {
+            return static_cast<bool>(*iface_ptr(*static_cast<const Wrap *>(this)));
+        }
     }
 
-    [[nodiscard]] std::type_index get_type_index() const final
+    // NOTE: these are part of the old callable interface, and they are not
+    // strictly needed as there are equivalent functions in tanuki. Consider removing
+    // them in the future.
+    auto get_type_index() const noexcept
     {
-        return typeid(T);
+        return value_type_index(*static_cast<const Wrap *>(this));
     }
+    template <typename T>
+    T *extract() noexcept
+    {
+        return value_ptr<T>(*static_cast<Wrap *>(this));
+    }
+    template <typename T>
+    const T *extract() const noexcept
+    {
+        return value_ptr<T>(*static_cast<const Wrap *>(this));
+    }
+};
 
-private:
-    // Serialization.
-    friend class boost::serialization::access;
-    template <typename Archive>
-    void serialize(Archive &ar, unsigned)
-    {
-        ar &boost::serialization::base_object<callable_inner_base<R, Args...>>(*this);
-        ar & m_value;
-    }
+template <typename R, typename... Args>
+struct HEYOKA_DLL_PUBLIC_INLINE_CLASS callable_ref_iface {
+    template <typename Wrap>
+    using type = callable_ref_iface_impl<Wrap, R, Args...>;
+};
+
+// Definition of the callable wrap.
+template <typename R, typename... Args>
+using callable_wrap_t = tanuki::wrap<callable_iface,
+                                     tanuki::config<R (*)(Args...), callable_ref_iface<R, Args...>::template type>{
+                                         // Similarly to std::function, ensure that callable can store
+                                         // in static storage pointers and reference wrappers.
+                                         // NOTE: reference wrappers are not guaranteed to have the size
+                                         // of a pointer, but in practice that should always be the case.
+                                         // In case this is a concern, static asserts can be added
+                                         // in the callable interface implementation.
+                                         .static_size = tanuki::holder_size<R (*)(Args...), callable_iface, R, Args...>,
+                                         .pointer_interface = false,
+                                         .explicit_generic_ctor = false},
+                                     R, Args...>;
+
+template <typename T>
+struct callable_impl {
+    static_assert(always_false_v<T>);
+};
+
+template <typename R, typename... Args>
+struct callable_impl<R(Args...)> {
+    using type = callable_wrap_t<R, Args...>;
 };
 
 } // namespace detail
 
 template <typename T>
-class HEYOKA_DLL_PUBLIC_INLINE_CLASS callable
-{
-    static_assert(detail::always_false_v<T>);
-};
+using callable = typename detail::callable_impl<T>::type;
 
 // Detect callable instances.
 template <typename>
@@ -117,188 +180,37 @@ struct is_any_callable : std::false_type {
 };
 
 template <typename R, typename... Args>
-struct is_any_callable<callable<R(Args...)>> : std::true_type {
+struct is_any_callable<detail::callable_wrap_t<R, Args...>> : std::true_type {
 };
-
-template <typename R, typename... Args>
-class HEYOKA_DLL_PUBLIC_INLINE_CLASS callable<R(Args...)>
-{
-    std::unique_ptr<detail::callable_inner_base<R, Args...>> m_ptr;
-
-    // Serialization.
-    friend class boost::serialization::access;
-    template <typename Archive>
-    void serialize(Archive &ar, unsigned)
-    {
-        ar & m_ptr;
-    }
-
-    // Detection of the candidate internal type from a generic T.
-    template <typename T>
-    using internal_type
-        = std::conditional_t<std::is_function_v<detail::uncvref_t<T>>, std::decay_t<T>, detail::uncvref_t<T>>;
-
-public:
-    // NOTE: default construction builds an empty callable.
-    callable() = default;
-    callable(const callable &other) : m_ptr(other ? other.m_ptr->clone() : nullptr) {}
-    callable(callable &&) noexcept = default;
-
-    // NOTE: generic ctor is enabled only if it does not
-    // compete with copy/move ctors.
-    template <typename T, std::enable_if_t<std::negation_v<std::is_same<callable, detail::uncvref_t<T>>>, int> = 0>
-    // NOLINTNEXTLINE(google-explicit-constructor, hicpp-explicit-conversions)
-    callable(T &&f)
-    {
-        using uT = detail::uncvref_t<T>;
-
-        // NOTE: if f is a nullptr or an empty callable or
-        // std::function, leave this empty.
-        if constexpr (detail::is_any_std_func_v<uT> || is_any_callable<uT>::value) {
-            if (!f) {
-                return;
-            }
-        }
-
-        if constexpr (std::is_pointer_v<uT> || std::is_member_object_pointer_v<uT>) {
-            if (f == nullptr) {
-                return;
-            }
-        }
-
-        m_ptr = std::make_unique<detail::callable_inner<internal_type<T &&>, R, Args...>>(std::forward<T>(f));
-    }
-
-    ~callable() = default;
-
-    callable &operator=(const callable &other)
-    {
-        if (this != &other) {
-            *this = callable(other);
-        }
-
-        return *this;
-    }
-    callable &operator=(callable &&) noexcept = default;
-
-    R operator()(Args... args) const
-    {
-        if (!m_ptr) {
-            throw std::bad_function_call();
-        }
-
-        return m_ptr->operator()(std::forward<Args>(args)...);
-    }
-
-    void swap(callable &other) noexcept
-    {
-        std::swap(m_ptr, other.m_ptr);
-    }
-
-    explicit operator bool() const noexcept
-    {
-        return static_cast<bool>(m_ptr);
-    }
-
-    [[nodiscard]] std::type_index get_type_index() const
-    {
-        if (m_ptr) {
-            return m_ptr->get_type_index();
-        } else {
-            return typeid(void);
-        }
-    }
-
-    // Extraction.
-    template <typename T>
-    const T *extract() const noexcept
-    {
-        if (!m_ptr) {
-            return nullptr;
-        }
-
-        auto p = dynamic_cast<const detail::callable_inner<T, R, Args...> *>(m_ptr.get());
-        return p == nullptr ? nullptr : &(p->m_value);
-    }
-    template <typename T>
-    T *extract() noexcept
-    {
-        if (!m_ptr) {
-            return nullptr;
-        }
-
-        auto p = dynamic_cast<detail::callable_inner<T, R, Args...> *>(m_ptr.get());
-        return p == nullptr ? nullptr : &(p->m_value);
-    }
-};
-
-template <typename R, typename... Args>
-inline void swap(callable<R(Args...)> &c0, callable<R(Args...)> &c1) noexcept
-{
-    c0.swap(c1);
-}
 
 HEYOKA_END_NAMESPACE
 
-// Disable Boost.Serialization tracking for the implementation details of callable.
-// NOTE: these bits are taken verbatim from the BOOST_CLASS_TRACKING macro, which does not support
-// class templates.
+#if defined(__GNUC__)
 
-namespace boost::serialization
-{
+#pragma GCC diagnostic pop
 
-template <typename R, typename... Args>
-struct tracking_level<heyoka::detail::callable_inner_base<R, Args...>> {
-    using tag = mpl::integral_c_tag;
-    using type = mpl::int_<track_never>;
-    BOOST_STATIC_CONSTANT(int, value = tracking_level::type::value);
-    BOOST_STATIC_ASSERT((mpl::greater<implementation_level<heyoka::detail::callable_inner_base<R, Args...>>,
-                                      mpl::int_<primitive_type>>::value));
-};
+#endif
 
-template <typename T, typename R, typename... Args>
-struct tracking_level<heyoka::detail::callable_inner<T, R, Args...>> {
-    using tag = mpl::integral_c_tag;
-    using type = mpl::int_<track_never>;
-    BOOST_STATIC_CONSTANT(int, value = tracking_level::type::value);
-    BOOST_STATIC_ASSERT((mpl::greater<implementation_level<heyoka::detail::callable_inner<T, R, Args...>>,
-                                      mpl::int_<primitive_type>>::value));
-};
+// Serialisation macros.
+// NOTE: by default, we build a custom name and pass it to TANUKI_S11N_WRAP_EXPORT_KEY2.
+// This allows us to reduce the size of the final guid wrt to what TANUKI_S11N_WRAP_EXPORT_KEY
+// would synthesise, and thus to ameliorate the "class name too long" issue.
+#define HEYOKA_S11N_CALLABLE_EXPORT_KEY(udc, ...)                                                                      \
+    TANUKI_S11N_WRAP_EXPORT_KEY2(udc, "heyoka::callable<" #__VA_ARGS__ ">@" #udc, heyoka::detail::callable_iface,      \
+                                 __VA_ARGS__)
 
-} // namespace boost::serialization
+#define HEYOKA_S11N_CALLABLE_EXPORT_KEY2(udc, gid, ...)                                                                \
+    TANUKI_S11N_WRAP_EXPORT_KEY2(udc, gid, heyoka::detail::callable_iface, __VA_ARGS__)
 
-// NOTE: these are verbatim re-implementations of the BOOST_CLASS_EXPORT_KEY
-// and BOOST_CLASS_EXPORT_IMPLEMENT macros, which do not work well with class templates.
-#define HEYOKA_S11N_CALLABLE_EXPORT_KEY(...)                                                                           \
-    namespace boost::serialization                                                                                     \
-    {                                                                                                                  \
-    template <>                                                                                                        \
-    struct guid_defined<heyoka::detail::callable_inner<__VA_ARGS__>> : boost::mpl::true_ {                             \
-    };                                                                                                                 \
-    template <>                                                                                                        \
-    inline const char *guid<heyoka::detail::callable_inner<__VA_ARGS__>>()                                             \
-    {                                                                                                                  \
-        /* NOTE: the stringize here will produce a name enclosed by brackets. */                                       \
-        return BOOST_PP_STRINGIZE((heyoka::detail::callable_inner<__VA_ARGS__>));                                      \
-    }                                                                                                                  \
-    }
+#define HEYOKA_S11N_CALLABLE_EXPORT_IMPLEMENT(udc, ...)                                                                \
+    TANUKI_S11N_WRAP_EXPORT_IMPLEMENT(udc, heyoka::detail::callable_iface, __VA_ARGS__)
 
-#define HEYOKA_S11N_CALLABLE_EXPORT_IMPLEMENT(...)                                                                     \
-    namespace boost::archive::detail::extra_detail                                                                     \
-    {                                                                                                                  \
-    template <>                                                                                                        \
-    struct init_guid<heyoka::detail::callable_inner<__VA_ARGS__>> {                                                    \
-        static guid_initializer<heyoka::detail::callable_inner<__VA_ARGS__>> const &g;                                 \
-    };                                                                                                                 \
-    guid_initializer<heyoka::detail::callable_inner<__VA_ARGS__>> const                                                \
-        &init_guid<heyoka::detail::callable_inner<__VA_ARGS__>>::g                                                     \
-        = ::boost::serialization::singleton<                                                                           \
-              guid_initializer<heyoka::detail::callable_inner<__VA_ARGS__>>>::get_mutable_instance()                   \
-              .export_guid();                                                                                          \
-    }
+#define HEYOKA_S11N_CALLABLE_EXPORT(udc, ...)                                                                          \
+    HEYOKA_S11N_CALLABLE_EXPORT_KEY(udc, __VA_ARGS__)                                                                  \
+    HEYOKA_S11N_CALLABLE_EXPORT_IMPLEMENT(udc, __VA_ARGS__)
 
-#define HEYOKA_S11N_CALLABLE_EXPORT(...)                                                                               \
-    HEYOKA_S11N_CALLABLE_EXPORT_KEY(__VA_ARGS__)                                                                       \
-    HEYOKA_S11N_CALLABLE_EXPORT_IMPLEMENT(__VA_ARGS__)
+#define HEYOKA_S11N_CALLABLE_EXPORT2(udc, gid, ...)                                                                    \
+    HEYOKA_S11N_CALLABLE_EXPORT_KEY2(udc, gid, __VA_ARGS__)                                                            \
+    HEYOKA_S11N_CALLABLE_EXPORT_IMPLEMENT(udc, __VA_ARGS__)
 
 #endif