+++ src/detail/event_detection.cpp
@@ -1126,35 +1126,30 @@ void taylor_adaptive<T>::ed_data::detect_events(const T &h, std::uint32_t order,
                 // Check if multiple roots are detected in the cooldown
                 // period for a terminal event. For non-terminal events,
                 // this will be unused.
-                [[maybe_unused]] const bool has_multi_roots = [&]() {
-                    if constexpr (detail::is_terminal_event_v<ev_type>) {
-                        // Establish the cooldown time.
-                        // NOTE: this is the same logic that is
-                        // employed in taylor.cpp to assign a cooldown
-                        // to a detected terminal event. g_eps has been checked
-                        // for finiteness early on, abs_der also has been checked for
-                        // finiteness above.
-                        const auto cd = (ev_vec[i].get_cooldown() >= 0)
-                                            ? ev_vec[i].get_cooldown()
-                                            : detail::taylor_deduce_cooldown(g_eps, abs_der);
+                // NOLINTNEXTLINE(misc-const-correctness)
+                [[maybe_unused]] bool has_multi_roots = false;
+                if constexpr (detail::is_terminal_event_v<ev_type>) {
+                    // Establish the cooldown time.
+                    // NOTE: this is the same logic that is
+                    // employed in taylor.cpp to assign a cooldown
+                    // to a detected terminal event. g_eps has been checked
+                    // for finiteness early on, abs_der also has been checked for
+                    // finiteness above.
+                    const auto cd = (ev_vec[i].get_cooldown() >= 0) ? ev_vec[i].get_cooldown()
+                                                                    : detail::taylor_deduce_cooldown(g_eps, abs_der);
 
-                        // NOTE: if the cooldown is zero, no sense to
-                        // run the check.
-                        if (cd == 0) {
-                            return false;
-                        }
-
+                    // NOTE: if the cooldown is zero, no sense to
+                    // run the check.
+                    if (cd != 0) {
                         // Evaluate the polynomial at the cooldown boundaries.
                         const auto e1 = detail::poly_eval(ptr, root + cd, order);
                         const auto e2 = detail::poly_eval(ptr, root - cd, order);
 
                         // We detect multiple roots within the cooldown
                         // if the signs of e1 and e2 are equal.
-                        return (e1 > 0) == (e2 > 0);
-                    } else {
-                        return false;
+                        has_multi_roots = (e1 > 0) == (e2 > 0);
                     }
-                }();
+                }
 
                 // Compute the sign of the derivative.
                 const auto d_sgn = detail::sgn(der);
@@ -1825,35 +1820,31 @@ void taylor_adaptive_batch<T>::ed_data::detect_events(const T *h_ptr, std::uint3
                     // Check if multiple roots are detected in the cooldown
                     // period for a terminal event. For non-terminal events,
                     // this will be unused.
-                    [[maybe_unused]] const bool has_multi_roots = [&]() {
-                        if constexpr (detail::is_terminal_event_v<ev_type>) {
-                            // Establish the cooldown time.
-                            // NOTE: this is the same logic that is
-                            // employed in taylor.cpp to assign a cooldown
-                            // to a detected terminal event. g_eps has been checked
-                            // for finiteness early on, abs_der also has been checked for
-                            // finiteness above.
-                            const auto cd = (ev_vec[i].get_cooldown() >= 0)
-                                                ? ev_vec[i].get_cooldown()
-                                                : detail::taylor_deduce_cooldown(g_eps, abs_der);
+                    // NOLINTNEXTLINE(misc-const-correctness)
+                    [[maybe_unused]] bool has_multi_roots = false;
+                    if constexpr (detail::is_terminal_event_v<ev_type>) {
+                        // Establish the cooldown time.
+                        // NOTE: this is the same logic that is
+                        // employed in taylor.cpp to assign a cooldown
+                        // to a detected terminal event. g_eps has been checked
+                        // for finiteness early on, abs_der also has been checked for
+                        // finiteness above.
+                        const auto cd = (ev_vec[i].get_cooldown() >= 0)
+                                            ? ev_vec[i].get_cooldown()
+                                            : detail::taylor_deduce_cooldown(g_eps, abs_der);
 
-                            // NOTE: if the cooldown is zero, no sense to
-                            // run the check.
-                            if (cd == 0) {
-                                return false;
-                            }
-
+                        // NOTE: if the cooldown is zero, no sense to
+                        // run the check.
+                        if (cd != 0) {
                             // Evaluate the polynomial at the cooldown boundaries.
                             const auto e1 = detail::poly_eval(ptr, root + cd, order);
                             const auto e2 = detail::poly_eval(ptr, root - cd, order);
 
                             // We detect multiple roots within the cooldown
                             // if the signs of e1 and e2 are equal.
-                            return (e1 > 0) == (e2 > 0);
-                        } else {
-                            return false;
+                            has_multi_roots = (e1 > 0) == (e2 > 0);
                         }
-                    }();
+                    }
 
                     // Compute sign of the derivative.
                     const auto d_sgn = detail::sgn(der);